@@ -28,14 +28,14 @@ const customerSchema = new mongoose.Schema({
     alternatePhone: String,
     email: String,
     panCard: String,
-    aadhaarCard: String,
+    aadharCard: String,
     utrNo: String,
     bookingType: String,
     location: String,
     address: String,
     district: String,
     taluka: String,
-    guntha: Number,
+    bookingArea: Number,
     rate: Number,
     plotArea: Number,
     actualAmount: Number,
